+++ lib/src/models/dm_room/dm_room.dart
@@ -1,4 +1,5 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
+import '../message/message.dart';
 
 part 'dm_room.freezed.dart';
 part 'dm_room.g.dart';
@@ -15,6 +16,9 @@ abstract class DmRoom with _$DmRoom {
     required String ts,
     required List<String> uids,
     String? topic,
+    String? name,
+    String? fname,
+    RocketChatMessage? lastMessage,
   }) = _DmRoom;
 
   factory DmRoom.fromJson(Map<String, dynamic> json) => _$DmRoomFromJson(json);